requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
idna==3.2
pyTelegramBotAPI==4.1.0
python-telegram-bot==13.7
pytz==2021.1
qrcode==7.3.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
telebot==0.0.4
tornado==6.1
tzdata==2021.2
tzlocal==3.0
urllib3==1.26.7
